feat: 클래스 이름 수정 및 메시저 기능 메서드 구현
